+++ plugins/phile/phpFastCache/Classes/Plugin.php
@@ -32,9 +32,20 @@ class Plugin extends AbstractPlugin
     {
         $storage = $this->settings['storage'];
         $config = $this->settings;
-        unset($config['active'], $config['storage']);
-        $psr16Cache = new \phpFastCache\Helper\Psr16Adapter($storage, $config);
 
+        if (!empty($config['phpFastCacheConfig'])) {
+            $cacheConfig = $config['phpFastCacheConfig'];
+        } else {
+            // Remove options unknown to Phpfastcache triggering errors
+            unset($config['active'], $config['storage']);
+            if ($storage == 'files') {
+                $cacheConfig = new \Phpfastcache\Drivers\Files\Config($config);
+            } else {
+                $cacheConfig = new \Phpfastcache\Config\ConfigurationOption($config);
+            }
+        }
+
+        $psr16Cache = new \Phpfastcache\Helper\Psr16Adapter($storage, $cacheConfig);
         $phileCache = new PhileToPsr16CacheAdapter($psr16Cache);
         ServiceLocator::registerService('Phile_Cache', $phileCache);
     }